chore: move the decorator for exact functions into utils
