+++ src/bridge/pipelines/protocols/map.py
@@ -2,7 +2,6 @@
 Abstract class for mapping between repository and metadata models.
 """
 
-import functools
 import importlib
 from abc import ABC, abstractmethod
 from enum import Enum
@@ -10,36 +9,6 @@ from functools import partial
 from typing import Any
 
 from pydantic import BaseModel, ConfigDict, field_serializer, field_validator
-
-
-def prepare_match_items(func: callable):
-    """
-    Check and normalize inputs for match methods.
-
-    Validation rule:
-    - None: returns 0 without further matching
-
-    Normalization rules:
-    - list/set/tuple: converted to set
-    # TODO: handle dicts?
-    """
-
-    @functools.wraps(func)
-    def wrapper(self, item1: Any, item2: Any) -> float:
-        # Missing inout value
-        if item1 is None or item2 is None:
-            return 0.0
-
-        def normalize(item):
-            if isinstance(item, (list, set, tuple)):
-                return set(item)
-            else:
-                return item
-
-        item1_norm, item2_norm = normalize(item1), normalize(item2)
-        return func(self, item1_norm, item2_norm)
-
-    return wrapper
 
 
 class Method(Enum):
