+++ src/bridge/pipelines/utils/__init__.py
@@ -1,0 +1,9 @@
+"""
+Utilities for pipelines.
+"""
+
+from .decorators import prepare_match_items
+
+__all__ = [
+    "prepare_match_items",
+]
